Aleksandrs K.: Lesson 11, level 4, Move FraudDetector and Test to fraud_detector_v1 package
